fix chatbot
