+++ lib/main.dart
@@ -8,7 +8,6 @@ import 'package:disappear/screens/challenge/detail_challenge_screen.dart';
 import 'package:disappear/screens/challenge/join_challenge_screen.dart';
 import 'package:disappear/screens/chatbot/chatbot_empty_screen.dart';
 import 'package:disappear/screens/chatbot/chatbot_screen.dart';
-import 'package:disappear/screens/chatbot/new_chatbot_screen.dart';
 import 'package:disappear/screens/address/add_new_address_screen.dart';
 import 'package:disappear/screens/checkout/address/checkout_address_screen.dart';
 import 'package:disappear/screens/address/address_list_screen.dart';
@@ -59,7 +58,7 @@ import 'package:disappear/view_models/auth/register/register_verification_view_m
 import 'package:disappear/view_models/auth/register/register_view_model.dart';
 import 'package:disappear/view_models/challenge_modules/challenge_main_view_model.dart';
 import 'package:disappear/view_models/cart/cart_view_model.dart';
-import 'package:disappear/view_models/chatbot/chatbot_viewmodel.dart';
+import 'package:disappear/view_models/chatbot/chatbot_view_model.dart';
 import 'package:disappear/view_models/checkout/checkout_address_view_model.dart';
 import 'package:disappear/view_models/checkout/checkout_payment_method_view_model.dart';
 import 'package:disappear/view_models/checkout/checkout_view_model.dart';
@@ -292,10 +291,8 @@ class App extends StatelessWidget {
         OrderListScreen.routePath: (context) => const OrderListScreen(),
         TelegramTransferScreen.routePath: (context) =>
             const TelegramTransferScreen(),
-        ChatBotScreen.routePath: (context) => const ChatBotScreen(),
+        ChatbotScreen.routePath: (context) => const ChatbotScreen(),
         ChatBotEmptyScreen.routePath: (context) => const ChatBotEmptyScreen(),
-        NewChatbotScreen.routePath: (context) => const NewChatbotScreen(),
-        ChatBotScreen.routePath: (context) => const ChatBotScreen(),
         ChangePasswordScreen.routePath: (context) => const ChangePasswordScreen(),
       },
       theme: ThemeData(
