+++ lib/screens/home/home_screen.dart
@@ -1,6 +1,6 @@
 import 'package:disappear/models/home/carousel_category_product_model.dart' as model;
 import 'package:disappear/models/home/challenge_article.dart';
-import 'package:disappear/screens/chatbot/new_chatbot_screen.dart';
+import 'package:disappear/screens/chatbot/chatbot_screen.dart';
 import 'package:disappear/screens/home/components/latest_articles.dart';
 import 'package:disappear/screens/home/components/latest_challenges.dart';
 import 'package:disappear/screens/home/components/placeholders/categories_placeholder.dart';
@@ -62,7 +62,7 @@ class _HomeScreenState extends State<HomeScreen> {
     return Scaffold(
       floatingActionButton: InkWell(
         onTap: () {
-          Navigator.pushNamed(context, NewChatbotScreen.routePath);
+          Navigator.pushNamed(context, ChatbotScreen.routePath);
         },
         child: Image.asset(
           'assets/img/Secretary.png',
